Docs: 컴포넌트, 렌더 폴더 따로 생성 후 분리
