+++ src/app.js
@@ -1,8 +1,9 @@
-import { renderDayTopContent, addDayTopTabEvent } from "./list.js";
+import { renderDayTopContent } from './render/dayTop.js';
+import { addDayTopTabEvent } from "./eventListener.js";
 
 function init() {
     const date = new Date();
-    const today = date.getDay() - 1;
+    let today = date.getDay() - 1;
     if(today === -1) today = 6;
     renderDayTopContent(today);
     addDayTopTabEvent();
